# Avoid head-of-line blocking in async spine.