--- a/crates/dbsp/src/trace/mod.rs
+++ b/crates/dbsp/src/trace/mod.rs
@@ -451,7 +451,7 @@ where
     type Builder: Builder<Self>;
 
     /// A type used to progressively merge batches.
-    type Merger: Merger<Self::Key, Self::Val, Self::Time, Self::R, Self>;
+    type Merger: Merger<Self::Key, Self::Val, Self::Time, Self::R, Self> + Send + Sync;
 
     /// Assemble an unordered vector of weighted items into a batch.
     #[allow(clippy::type_complexity)]