This means we work on multiple merges simultaneously. We can re-use the fuel approach to do that. But we need to make sure we can call the work-closure multiple times to continue merging. This involved being able to share Batch::Merger across threads too (if there is a better way instead of having the Send+Sync bound there I'd be happy to avoid that).