@gz 2024-06-20
Some numbers using `max_storage_rows = 10_000`:


- This PR (avoid head-of-line-blocking): TMPDIR=/media/gz/benchmarks cargo bench --bench nexmark -- --first-event-rate=10000000 --max-events=50000000 --cpu-cores 16 --num-event-generators 16 --source-buffer-size 10000 --input-batch-size 40000 --query q9 --storage 10000

┌───────┬────────────┬───────┬─────────┬─────────────────┬──────────────────┬───────────────┬───────────────┬───────────┬─────────────┬─────────┬────────────┬───────────┬───────────┬───────────┬────────────────┬───────┬─────────────┬────────────────┐
│ Query │ #Events    │ Cores │ Elapsed │ Cores * Elapsed │ Throughput/Cores │ Total Usr CPU │ Total Sys CPU │ Peak RSS  │ Page Faults │ # Files │ Avg WrSz   │ Avg RdSz  │ Writes    │ Reads     │ Cache Hit Rate │ Cpcts │ Cpct Saving │ Cpct Stall [s] │
├───────┼────────────┼───────┼─────────┼─────────────────┼──────────────────┼───────────────┼───────────────┼───────────┼─────────────┼─────────┼────────────┼───────────┼───────────┼───────────┼────────────────┼───────┼─────────────┼────────────────┤
│ Q9    │ 50,000,000 │ 16    │ 94.707s │ 1515.312s       │ 32.996 K/s       │ 1519.828s     │ 770.478s      │ 11.72 GiB │ 7           │ 14580   │ 948.36 KiB │ 16.40 KiB │ 76.86 GiB │ 66.60 GiB │ 98%            │ 78058 │ 92426       │ 0              │
└───────┴────────────┴───────┴─────────┴─────────────────┴──────────────────┴───────────────┴───────────────┴───────────┴─────────────┴─────────┴────────────┴───────────┴───────────┴───────────┴────────────────┴───────┴─────────────┴────────────────┘

- current master: TMPDIR=/media/gz/benchmarks cargo bench --bench nexmark -- --first-event-rate=10000000 --max-events=50000000 --cpu-cores 16 --num-event-generators 16 --source-buffer-size 10000 --input-batch-size 40000 --query q9 --storage 10000

┌───────┬────────────┬───────┬──────────┬────────────────┬────────────────┬───────────────┬───────────────┬───────────┬─────────────┬─────────┬────────────┬───────────┬────────────┬────────────┬────────────────┬───────┬─────────────┬────────────────┐
│ Query │ #Events    │ Cores │ Elapsed  │ Cores * Elaps+ │ Throughput/Co+ │ Total Usr CPU │ Total Sys CPU │ Peak RSS  │ Page Faults │ # Files │ Avg WrSz   │ Avg RdSz  │ Writes     │ Reads      │ Cache Hit Rate │ Cpcts │ Cpct Saving │ Cpct Stall [s] │
├───────┼────────────┼───────┼──────────┼────────────────┼────────────────┼───────────────┼───────────────┼───────────┼─────────────┼─────────┼────────────┼───────────┼────────────┼────────────┼────────────────┼───────┼─────────────┼────────────────┤
│ Q9    │ 50,000,000 │ 16    │ 179.502s │ 2872.040s      │ 17.409 K/s     │ 1459.758s     │ 1625.908s     │ 13.36 GiB │ 3           │ 13920   │ 974.52 KiB │ 11.84 KiB │ 123.20 GiB │ 619.58 GiB │ 78%            │ 77361 │ 88098       │ 123            │
└───────┴────────────┴───────┴──────────┴────────────────┴────────────────┴───────────────┴───────────────┴───────────┴─────────────┴─────────┴────────────┴───────────┴────────────┴────────────┴────────────────┴───────┴─────────────┴────────────────┘


- old spine fueled: TMPDIR=/media/gz/benchmarks cargo bench --bench nexmark -- --first-event-rate=10000000 --max-events=50000000 --cpu-cores 16 --num-event-generators 16 --source-buffer-size 10000 --input-batch-size 40000 --query q9 --storage 10000

┌───────┬───────────┬───────┬──────────┬───────────┬───────────┬──────────┬──────────┬──────────┬──────────┬──────────┬──────────┬──────────┬─────────┬──────────┬──────────┬──────────┬──────────┬──────────┬──────────┬──────────┬──────────┬──────────┐
│ Query │ #Events   │ Cores │ Elapsed  │ Cores * + │ Throughp+ │ Total U+ │ Total S+ │ Current+ │ Peak RSS │ Current+ │ Peak Co+ │ Page Fa+ │ # Files │ # Writes │ # Reads  │ Avg WrSz │ Avg RdSz │ Writes   │ Reads    │ Cache H+ │ Cache M+ │ Compact+ │
├───────┼───────────┼───────┼──────────┼───────────┼───────────┼──────────┼──────────┼──────────┼──────────┼──────────┼──────────┼──────────┼─────────┼──────────┼──────────┼──────────┼──────────┼──────────┼──────────┼──────────┼──────────┼──────────┤
│ Q9    │ 50,000,0+ │ 16    │ 174.652s │ 2794.432s │ 17.893 K+ │ 1028.90+ │ 666.628s │ 10.45 G+ │ 8.97 GiB │ 10.45 G+ │ 10.45 G+ │ 10       │ 14273   │ 108292   │ 45802594 │ 960.18 + │ 15.04 K+ │ 99.16 G+ │ 656.96 + │ 1394048+ │ 45802594 │ 79618    │
└───────┴───────────┴───────┴──────────┴───────────┴───────────┴──────────┴──────────┴──────────┴──────────┴──────────┴──────────┴──────────┴─────────┴──────────┴──────────┴──────────┴──────────┴──────────┴──────────┴──────────┴──────────┴──────────┘