@mihaibudiu 2024-06-20
these tables are confusing because the headers change. can you tell it not to truncate the headers?