@mihaibudiu 2024-06-20
I hope that when you measure the time you have completed all pending operations