@gz 2024-06-20
>you have completed all pending operations

what is "all pending operations"?