@mihaibudiu 2024-06-20
everything that had to be flushed has been flushed