@blp 2024-06-20
> This involved being able to share Batch::Merger across threads too (if there is a better way instead of having the Send+Sync bound there I'd be happy to avoid that).

Why does the merger need to be shared across threads? Off-hand, I'd expect that the merger could be created, executed, and finalized in the background thread, without ever being moved between threads.