@blp 2024-06-20
> > Why does the merger need to be shared across threads? Off-hand, I'd expect that the merger could be created, executed, and finalized in the background thread, without ever being moved between threads.
> 
> It doesn't really. But I couldn't figure out how to move this variable into the closure itself (so I can call it repeatedly with fuel) if I can do that, it's no longer necessary.
> 
> https://github.com/feldera/feldera/pull/1912/files#diff-68a0dc26167f0169660b37db7633b90b3265cffbeb407c3acc1269c8cd544303R587
> 
> hm now that I think more about it maybe there is a way with take/swap etc. I'll give it a try

Oh I see the issue. I think there must be a good solution (but I thought for 5 minutes and didn't spot one).