@lalithsuresh 2024-06-20
One thing to pay close attention to with all storage tests, is the ratio of pipeline workers vs threads you run vs available cores. I don't think we automatically pick the number of workers for the pipeline.

On your own machines, you might have way more cores than workers we use by default for example, so the extra work done by additional background threads doesn't matter. When running on modest VMs in the cloud (8-16 cores), we might not have that kind of breathing room.